Gymnasium env loop (#768)
Initial work to enable agent training on Atari environments. Currently
supporting the Gymnasium API, however we can consider swapping for a
more performant API in the future.

Main changes:
- Move nested functions in core.py to private functions instead. The
reason for this is because in gymnasium_loop.py, we end up tracing the
public functions multiple times, which re-creates the jitted function
every time. So, this prevents getting compile cache misses
 - Gymnasium_loop is similar to gymnax_loop except for it doesn't support on policy training
- The environment loop cycle is not jitted, which means that after every
CPU environment step, we need to move the observations to the device.